eliminando factura

+++ Backend/src/controllers/delete.controllers.js
@@ -138,15 +138,21 @@ export const deleteServiceInventory = async (req, res) => {
 export const deleteServiceCustomer = async (req, res) => {
   try {
     const { identificacion } = req.params;
-    const [row] = await pool.query(
-      "DELETE FROM servicio_cliente WHERE identificacion = ? ",
-      [identificacion]
-    );
-    if (row.affectedRows === 0) {
-      return res.status(404).json({
-        message: "No se encontró el servicio.",
-      });
+    const [rows] = await pool.query("", [identificacion])
+    if(rows.length>0){
+      return res.status(404).json({message:"No se puede eliminar el registro porque tiene una factura"})
+    }else{
+      const [row] = await pool.query(
+        "DELETE FROM servicio_cliente WHERE identificacion = ? ",
+        [identificacion]
+      );
+      if (row.affectedRows === 0) {
+        return res.status(404).json({
+          message: "No se encontró el servicio.",
+        });
+      }
     }
+
     res.send({
       message: "Servicio eliminado correctamente",
       identificacion,
